Refactor RFQ Creation and Management for Enhanced Pricing and Currency Handling
- Renamed price field to unitPrice in the RFQ creation API to unify pricing terminology.
- Updated pricing calculations to ensure suggested unit price is derived from inventory cost when not provided.
- Enhanced total budget calculation in RFQ detail view to accurately reflect unit prices based on selected currency.
- Removed deprecated layout and email parser components to streamline RFQ management interface.
- Integrated currency conversion across various components to ensure consistent pricing display in CAD and USD.
