--- app/rfq-management/new/email-parser/page.tsx
@@ -1,12 +1,0 @@
-"use client"
-
-import { TabsContent } from "@/components/ui/tabs"
-import EmailParser from "@/components/email-parser"
-
-export default function EmailParserPage() {
-  return (
-    <TabsContent value="email" forceMount>
-      <EmailParser />
-    </TabsContent>
-  )
-}
